+++ Modules/CIPPCore/Public/Standards/Invoke-CIPPStandardPhishingSimulations.ps1
@@ -16,7 +16,7 @@ function Invoke-CIPPStandardPhishingSimulations {
         ADDEDCOMPONENT
             {"type":"autoComplete","multiple":true,"creatable":true,"required":true,"label":"Phishing Simulation Domains","name":"standards.PhishingSimulations.Domains"}
             {"type":"autoComplete","multiple":true,"creatable":true,"required":true,"label":"Phishing Simulation Sender IP Ranges","name":"standards.PhishingSimulations.SenderIpRanges"}
-            {"type":"autoComplete","multiple":true,"creatable":true,"required":true,"label":"Phishing Simulation Urls","name":"standards.PhishingSimulations.PhishingSimUrls"}
+            {"type":"autoComplete","multiple":true,"creatable":true,"required":false,"label":"Phishing Simulation Urls","name":"standards.PhishingSimulations.PhishingSimUrls"}
         IMPACT
             Medium Impact
         ADDEDDATE
@@ -45,10 +45,18 @@ function Invoke-CIPPStandardPhishingSimulations {
     Select-Object -Property Identity,Name,SenderIpRanges,Domains,SenderDomainIs
 
     [String[]]$AddSenderIpRanges = $Settings.SenderIpRanges.value | Where-Object { $_ -notin $RuleState.SenderIpRanges }
-    [String[]]$RemoveSenderIpRanges = $RuleState.SenderIpRanges | Where-Object { $_ -notin $Settings.SenderIpRanges.value }
+    if ($Settings.RemoveExtraUrls -eq $true) {
+        [String[]]$RemoveSenderIpRanges = $RuleState.SenderIpRanges | Where-Object { $_ -notin $Settings.SenderIpRanges.value }
+    } else {
+        $RemoveSenderIpRanges = @()
+    }
 
     [String[]]$AddDomains = $Settings.Domains.value | Where-Object { $_ -notin $RuleState.Domains }
-    [String[]]$RemoveDomains = $RuleState.Domains | Where-Object { $_ -notin $Settings.Domains.value }
+    if ($Settings.RemoveExtraUrls -eq $true) {
+        [String[]]$RemoveDomains = $RuleState.Domains | Where-Object { $_ -notin $Settings.Domains.value }
+    } else {
+        $RemoveDomains = @()
+    }
 
     $RuleIsCorrect = ($RuleState.Name -like "*PhishSimOverr*") -and
     ($AddSenderIpRanges.Count -eq 0 -and $RemoveSenderIpRanges.Count -eq 0) -and
@@ -59,18 +67,28 @@ function Invoke-CIPPStandardPhishingSimulations {
     Select-Object -Property Value
 
     [String[]]$AddEntries = $Settings.PhishingSimUrls.value | Where-Object { $_ -notin $SimUrlState.value }
-    [String[]]$RemoveEntries = $SimUrlState.value | Where-Object { $_ -notin $Settings.PhishingSimUrls.value }
+    if ($Settings.RemoveExtraUrls -eq $true) {
+        [String[]]$RemoveEntries = $SimUrlState.value | Where-Object { $_ -notin $Settings.PhishingSimUrls.value }
+    } else {
+        $RemoveEntries = @()
+    }
 
     $PhishingSimUrlsIsCorrect = ($AddEntries.Count -eq 0 -and $RemoveEntries.Count -eq 0)
 
     # Check state for all components
     $StateIsCorrect = $PolicyIsCorrect -and $RuleIsCorrect -and $PhishingSimUrlsIsCorrect
 
+    $CompareField = [PSCustomObject]@{
+        Domains         = $RuleState.Domains -join ', '
+        SenderIpRanges  = $RuleState.SenderIpRanges -join ', '
+        PhishingSimUrls = $SimUrlState.value -join ', '
+    }
+
     If ($Settings.remediate -eq $true) {
         If ($StateIsCorrect -eq $true) {
             Write-LogMessage -API 'Standards' -Tenant $Tenant -message 'Advanced Phishing Simulations already correctly configured' -sev Info
         } Else {
-            # Remedidate incorrect Phishing Simulations Policy
+            # Remediate incorrect Phishing Simulations Policy
             If ($PolicyIsCorrect -eq $false) {
                 If ($PolicyState.Name -eq 'PhishSimOverridePolicy') {
                     Try {
@@ -89,7 +107,7 @@ function Invoke-CIPPStandardPhishingSimulations {
                 }
             }
 
-            # Remedidate incorrect Phishing Simulations Policy Rule
+            # Remediate incorrect Phishing Simulations Policy Rule
             If ($RuleIsCorrect -eq $false) {
                 If ($RuleState.Name -like "*PhishSimOverr*") {
                     $cmdParams = @{
@@ -121,20 +139,22 @@ function Invoke-CIPPStandardPhishingSimulations {
                 }
             }
 
-            # Remedidate incorrect Phishing Simulations URLs
+            # Remediate incorrect Phishing Simulations URLs
             If ($PhishingSimUrlsIsCorrect -eq $false) {
                 $cmdParams = @{
                     ListType = 'Url'
                     ListSubType = 'AdvancedDelivery'
                 }
-                # Remove entries that are not in the settings
-                If ($RemoveEntries.Count -gt 0) {
-                    $cmdParams.Entries = $RemoveEntries
-                    Try {
-                        $null = New-ExoRequest -TenantId $Tenant -cmdlet 'Remove-TenantAllowBlockListItems' -cmdParams $cmdParams
-                        Write-LogMessage -API 'Standards' -Tenant $Tenant -message "Removed Phishing Simulation URLs from Allowlist." -sev Info
-                    } Catch {
-                        Write-LogMessage -API 'Standards' -Tenant $Tenant -message "Failed to remove Phishing Simulation URLs from Allowlist." -sev Error -LogData $_
+                if ($Settings.RemoveExtraUrls -eq $true) {
+                    # Remove entries that are not in the settings
+                    If ($RemoveEntries.Count -gt 0) {
+                        $cmdParams.Entries = $RemoveEntries
+                        Try {
+                            $null = New-ExoRequest -TenantId $Tenant -cmdlet 'Remove-TenantAllowBlockListItems' -cmdParams $cmdParams
+                            Write-LogMessage -API 'Standards' -Tenant $Tenant -message "Removed Phishing Simulation URLs from Allowlist." -sev Info
+                        } Catch {
+                            Write-LogMessage -API 'Standards' -Tenant $Tenant -message "Failed to remove Phishing Simulation URLs from Allowlist." -sev Error -LogData $_
+                        }
                     }
                 }
                 # Add entries that are in the settings
@@ -157,18 +177,14 @@ function Invoke-CIPPStandardPhishingSimulations {
         If ($StateIsCorrect -eq $true) {
             Write-LogMessage -API 'Standards' -Tenant $Tenant -message 'Phishing Simulation Configuration is correctly configured' -sev Info
         } Else {
-            Write-StandardsAlert -message 'Phishing Simulation Configuration is not correctly configured' -object $CurrentState -tenant $Tenant -standardName 'PhishingSimulations' -standardId $Settings.standardId
+            Write-StandardsAlert -message 'Phishing Simulation Configuration is not correctly configured' -object $CompareField -tenant $Tenant -standardName 'PhishingSimulations' -standardId $Settings.standardId
             Write-LogMessage -API 'Standards' -Tenant $Tenant -message 'Phishing Simulation Configuration is not correctly configured' -sev Info
         }
     }
 
     If ($Settings.report -eq $true) {
+        $FieldValue = $StateIsCorrect ? $true : $CompareField
         Add-CIPPBPAField -FieldName 'PhishingSimulations' -FieldValue $StateIsCorrect -StoreAs bool -Tenant $Tenant
-        If ($StateIsCorrect -eq $true) {
-            $FieldValue = $true
-        } Else {
-            $FieldValue = $CurrentState ? $CurrentState : $false
-        }
         Set-CIPPStandardsCompareField -FieldName 'standards.PhishingSimulations' -FieldValue $FieldValue -Tenant $Tenant
     }
 }
